Attempt to fix #613

+++ test/bash/watch.sh
@@ -5,33 +5,45 @@ source "${SRC}/include.sh"
 
 cd $file_path
 
+function waituntil {
+  for (( i = 0; i <= $1; i++ )); do
+      sleep 0.2
+      echo -n "."
+  done
+  echo ""
+}
+
 echo -e "\033[1mRunning tests:\033[0m"
 
 #####################
 # Watch for changes #
 #####################
+if [ -f server-watch.js ]; then
+  rm server-watch.js
+fi
+
 $pm2 kill
+
+sleep 1 
 
 cp server-watch.bak.js server-watch.js
 
 $pm2 start --watch server-watch.js
 
-sleep 1
+should 'process should be watched' 'watch: true' 1
 
-echo "setTimeout(function() { console.log('still ok') }, 200)" >> server-watch.js
-
-for (( i = 0; i <= 10; i++ )); do
-    sleep 0.2
-    echo -n "."
-done
+echo "console.log('test');" >> server-watch.js
 
 $pm2 list
 
 should 'process should have been restarted' 'restart_time: 1' 1
+should 'process should be online' "status: 'online'" 1
 
 $pm2 kill
-rm server-watch.js
 
+sleep 1
+
+rm server-watch.js
 
 ###############
 
@@ -72,8 +84,6 @@ cp server-watch.bak.js server-watch.js
 
 $pm2 start --watch server-watch.js
 
-sleep 1
-
 $pm2 restart 0
 
 should 'process should be watched' 'watch: true' 1
@@ -82,18 +92,7 @@ $pm2 stop --watch 0
 
 should 'process should have stopped beeing watched' 'watch: false' 1
 
-$pm2 list
-
 echo "setInterval(function() { console.log('still ok'); }, 100);" > server-watch.js
-
-cat server-watch.js
-
-for (( i = 0; i <= 10; i++ )); do
-    sleep 0.1
-    echo -n "."
-done
-
-$pm2 list
 
 should 'process should not have been restarted on file change' 'restart_time: 1' 1
 
@@ -104,32 +103,19 @@ $pm2 restart 0
 should 'process should restart and not be watched' 'watch: false' 1
 
 $pm2 restart --watch 0
-
-should 'process should restart and be watched' 'watch: true' 1
-
-sleep 1
-
-echo "setTimeout(function() { console.log('watch me!') })" >> server-watch.js
-
-for (( i = 0; i <= 10; i++ )); do
-    sleep 0.2
-    echo -n "."
-done
-
-$pm2 list
-
-should 'process should have restart because of a file change' 'restart_time: 4' 1
+should 'process should be watched' 'watch: true' 1
 
 $pm2 kill
-rm server-watch.js
+sleep 1
 
+rm server-watch.js
 #############
 # JSON test #
 #############
 # we've already seen before that "watch: true" is really watching when changing a file
 
 $pm2 start --watch all.json
-sleep 1
+
 should 'processes should be watched' 'watch: true' 8
 
 $pm2 stop --watch all
@@ -140,3 +126,22 @@ $pm2 restart --watch all
 should 'processes should be watched' 'watch: true' 8
 
 $pm2 kill
+
+##########
+# delete #
+##########
+
+cp server-watch.bak.js server-watch.js
+
+$pm2 start server-watch.js --watch 
+$pm2 stop 0
+$pm2 delete 0
+
+echo "setTimeout(function() { console.log('watch me!') })" >> server-watch.js
+
+waituntil 10
+
+should 'process should not have been restarted' 'watch: true' 0
+
+$pm2 kill
+rm server-watch.js
